mods/requirements.txt
tqdm
kociemba
Pillow
matplotlib
git+https://github.com/pybluez/pybluez.git#egg=pybluez
pyjnius